+++ incapsula/resource_site.go
@@ -58,6 +58,7 @@ func resourceSite() *schema.Resource {
 				Description: "Manually set the web server IP/CNAME.",
 				Type:        schema.TypeString,
 				Optional:    true,
+				Computed:    true,
 			},
 			"force_ssl": {
 				Description: "If this value is true, manually set the site to support SSL. This option is only available for sites with manually configured IP/CNAME and for specific accounts.",
@@ -270,14 +271,39 @@ func resourceSite() *schema.Resource {
 				Computed:    true,
 				Optional:    true,
 			},
+			"txt_record_value_one": {
+				Description: "Create or modify a TXT records defined for the site in Cloud WAF.",
+				Type:        schema.TypeString,
+				Optional:    true,
+			},
+			"txt_record_value_two": {
+				Description: "Create or modify a TXT records defined for the site in Cloud WAF.",
+				Type:        schema.TypeString,
+				Optional:    true,
+			},
+			"txt_record_value_three": {
+				Description: "Create or modify a TXT records defined for the site in Cloud WAF.",
+				Type:        schema.TypeString,
+				Optional:    true,
+			},
+			"txt_record_value_four": {
+				Description: "Create or modify a TXT records defined for the site in Cloud WAF.",
+				Type:        schema.TypeString,
+				Optional:    true,
+			},
+			"txt_record_value_five": {
+				Description: "Create or modify a TXT records defined for the site in Cloud WAF.",
+				Type:        schema.TypeString,
+				Optional:    true,
+			},
 			"naked_domain_san": {
 				Description: "Use 'true' to add the naked domain SAN to a www site’s SSL certificate. Default value: true",
-				Type:        schema.TypeString,
+				Type:        schema.TypeBool,
 				Optional:    true,
 			},
 			"wildcard_san": {
 				Description: "Use 'true' to add the wildcard SAN or 'false' to add the full domain SAN to the site’s SSL certificate. Default value: true",
-				Type:        schema.TypeString,
+				Type:        schema.TypeBool,
 				Optional:    true,
 			},
 			// Computed Attributes
@@ -341,8 +367,8 @@ func resourceSiteCreate(d *schema.ResourceData, m interface{}) error {
 		d.Get("site_ip").(string),
 		d.Get("force_ssl").(string),
 		d.Get("account_id").(int),
-		d.Get("naked_domain_san").(string),
-		d.Get("wildcard_san").(string),
+		d.Get("naked_domain_san").(bool),
+		d.Get("wildcard_san").(bool),
 	)
 
 	if err != nil {
@@ -379,6 +405,11 @@ func resourceSiteCreate(d *schema.ResourceData, m interface{}) error {
 	}
 
 	err = updatePerformanceSettings(client, d)
+	if err != nil {
+		return err
+	}
+
+	err = updateTXTRecords(client, d)
 	if err != nil {
 		return err
 	}
@@ -473,6 +504,18 @@ func resourceSiteRead(d *schema.ResourceData, m interface{}) error {
 	d.Set("hashing_enabled", maskingResponse.HashingEnabled)
 	d.Set("hash_salt", maskingResponse.HashSalt)
 
+	// Get the TXT records for the site
+	txtResponse, err := client.GetTXTRecords(d.Id())
+	if err != nil {
+		log.Printf("[ERROR] Could not read Incapsula site TXT records for domain: %s and site id: %d, %s\n", domain, siteID, err)
+		return err
+	}
+	d.Set("txt_record_value_one", txtResponse.TxtRecordValueOne)
+	d.Set("txt_record_value_two", txtResponse.TxtRecordValueTwo)
+	d.Set("txt_record_value_three", txtResponse.TxtRecordValueThree)
+	d.Set("txt_record_value_four", txtResponse.TxtRecordValueFour)
+	d.Set("txt_record_value_five", txtResponse.TxtRecordValueFive)
+
 	// Get the performance settings for the site
 	performanceSettingsResponse, _, err := client.GetPerformanceSettings(d.Id())
 	if err != nil {
@@ -548,6 +591,11 @@ func resourceSiteUpdate(d *schema.ResourceData, m interface{}) error {
 		return err
 	}
 
+	err = updateTXTRecords(client, d)
+	if err != nil {
+		return err
+	}
+
 	// Set the rest of the state from the resource read
 	return resourceSiteRead(d, m)
 }
@@ -611,6 +659,23 @@ func updateMaskingSettings(client *Client, d *schema.ResourceData) error {
 		err := client.UpdateMaskingSettings(d.Id(), &maskingSettings)
 		if err != nil {
 			log.Printf("[ERROR] Could not update Incapsula site masking settings for site_id: %s %s\n", d.Id(), err)
+			return err
+		}
+	}
+	return nil
+}
+
+func updateTXTRecords(client *Client, d *schema.ResourceData) error {
+	if d.HasChange("txt_record_value_one") ||
+		d.HasChange("txt_record_value_two") ||
+		d.HasChange("txt_record_value_three") ||
+		d.HasChange("txt_record_value_four") ||
+		d.HasChange("txt_record_value_five") {
+
+		_, err := client.UpdateTXTRecord(d.Id(), d.Get("txt_record_value_one").(string), d.Get("txt_record_value_two").(string),
+			d.Get("txt_record_value_three").(string), d.Get("txt_record_value_four").(string), d.Get("txt_record_value_five").(string))
+		if err != nil {
+			log.Printf("[ERROR] Could not update Incapsula site TXT record(s) for site_id: %s %s\n", d.Id(), err)
 			return err
 		}
 	}